fix layout page
