+++ src/components/dashboard.tsx
@@ -1,0 +1,13 @@
+"use client";
+import React from "react";
+// import "./style/dashboard.css";
+const Dashboard: React.FC = () => {
+  return (
+    <div className="al">
+      <h2>داشبورد</h2>
+      <p>محتوای داشبورد اینجا نمایش داده می‌شود.</p>
+    </div>
+  );
+};
+
+export default Dashboard;
