+++ src/components/style/dashbord.css
@@ -1,0 +1,4 @@
+.al{
+    display: flex;
+    flex-direction: row;
+}
