+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 // import App from './App';
 import reportWebVitals from './reportWebVitals';
 import 'antd/dist/reset.css'; // برای ریست کردن استایل
-import App from './app/page.tsx';
+import App from './app/home-page.tsx';
